Rename qRegisterType to qTypeId
Registration concept is practically gone. The id is just a pointer
so there is no global registry nor saving state.

Some extensions are allowed to execute code, but that should be
discouraged. Sadly it seems that it is not always possible to avoid
it, so the functionality needs to stay.