Change TypeIdData
The structure changed from function pointer to extensions dispatcher.
That means that typeId is not longer a function pointer, but a pointer
to a structure. Currently the dispatcher is a std::unordered_map, but
it should be optimize out later. So in the end we can have the structure
that function and data pointers for every extension.

The new layout supports runtime defined types much better. Sadly it
conceptually adds one additional de-reference when accessing
the type data. Hopefully it the structure would fit cache line so
there would be not visible performance costs, that should happen
at least for the first extension.

This patch could use more compile time structures to avoid runtime
overhead. In particular usage of the std::unordered_map is odd, but
as we know that it is possible to write something better there, we
can delegate it to a later patch.

As the template code became smarter the patch had to de-couple
Name_dlsym and Name_hash. These classes should not be in class
hierarchy anyway.