Kill extensions tagging
The tag was not different from type id, by replacing it by the id
we gain a bit of functionality. Once we do not have to store two
different kinds of ids, second it means that we can register
extensions to extensions and we got names handling for free.

The costs are minimal, as the tag was already based on static
initialization, so I do not expect performance differences. To break
infinite recursion one ugliness had to be injected (specialization
of qTypeId for Name_dlsym).