Store initial extensions list just after TypeIdData
We do not expect huge list of extensions, in the worst case maybe
around 10. So linear search should be the fastest.

By storing the extensions together with the TypeIdData we have big
chances to improve cache hits rate and we reduce amount of
neccessery allocations.

In long term we should be able to get rid of the unordered_map from
TypeIdData, by using extension list.