Remove invalid TODO
4805f95d0bddca7b2a0a3a682cca1dd2a785a0c9 ("Kill extensions tagging")
removed need for alignment. Before an operation and an extension tag
where kept together therefore we needed to ensure enough room at
the end of the tag (which was a function pointer). The new system,
because of runtime types, can not use this optimization as the types
require external data, so the tag, can not be just a function
pointer.