Automatically handle default extensions
Some types are not passing requirements of certain metatype
extensions. For example in Qt5 all types needed to be default
and copy constructible, but the requirement came from QVariant
not from QMetaType. In Qt6 it make sense to lift it. So if a user
calls:
  qTypeId<Foo>();
Qt would try to create all default extensions describing Foo type.
All extensions that can not work with the type will be silently
skipped. If an extension is explicitly created for a type that
can not be supported, compilation would fail.