Reorganize the code
This patch extracts library code and a manual test form the example.
This would allow to write more auto tests without rebuilding
everything over and over.

Use cmake instead of qmake, just for learning purposes. The result
works, but it is probably not that clean as it should. In particular
tests needs explicitly specify include path to the library.